requirements.txt
qrcode[pil]>=7.4.2
pillow>=10.0.0
flask>=2.3.0